mime_types: Move INLINE_MIME_TYPES to prevent future import loops.
