+++ zerver/lib/upload/s3.py
@@ -14,9 +14,10 @@ from django.conf import settings
 from django.utils.http import content_disposition_header
 from typing_extensions import override
 
+from zerver.lib.mime_types import INLINE_MIME_TYPES
 from zerver.lib.partial import partial
 from zerver.lib.thumbnail import resize_logo, resize_realm_icon
-from zerver.lib.upload.base import INLINE_MIME_TYPES, StreamingSourceWithSize, ZulipUploadBackend
+from zerver.lib.upload.base import StreamingSourceWithSize, ZulipUploadBackend
 from zerver.models import Realm, RealmEmoji, UserProfile
 
 if TYPE_CHECKING:
